+++ cnccode/controller_classes/CTCustomer.inc.php
@@ -12,6 +12,13 @@ require_once($cfg['path_bu'] . '/BUProject.inc.php');
 require_once($cfg['path_bu'] . '/BUSector.inc.php');
 require_once($cfg['path_dbe'] . '/DBEJOrdhead.inc.php');
 require_once($cfg['path_bu'] . '/BUPortalCustomerDocument.inc.php');
+require_once($cfg["path_bu"] . "/BURenBroadband.inc.php");
+require_once($cfg["path_bu"] . "/BURenContract.inc.php");
+require_once($cfg["path_bu"] . "/BURenQuotation.inc.php");
+require_once($cfg["path_bu"] . "/BURenDomain.inc.php");
+require_once($cfg["path_bu"] . "/BURenHosting.inc.php");
+require_once($cfg["path_bu"] . "/BUExternalItem.inc.php");
+require_once($cfg["path_bu"] . "/BUCustomerItem.inc.php");
 require_once($cfg['path_ct'] . '/CTCNC.inc.php');
 // Parameters
 define(
@@ -1589,6 +1596,48 @@ ORDER BY cus_name ASC  ";
         }
     }
 
+
+    private function extractValidContracts($something)
+    {
+
+        $contracts = [];
+        $validItems = [
+            "2nd Site",
+            "Internet Services",
+            "Managed Service",
+            "ServerCare",
+            "ServiceDesk",
+            "Telecom Services",
+            "PrePay"
+        ];
+
+        while ($something->fetchNext()) {
+
+            $continue = true;
+
+            foreach ($validItems as $item) {
+                if (strpos(
+                        $something->getValue('itemTypeDescription'),
+                        $item
+                    ) !== false) {
+                    $continue = false;
+                }
+            }
+
+            if ($continue) {
+                continue;
+            }
+            $contracts[] = [
+                'itemTypeDescription' => $something->getValue("itemTypeDescription"),
+                'customerItemID'      => $something->getValue("customerItemID"),
+                'itemDescription'     => $something->getValue('itemDescription')
+            ];
+        }
+
+        return $contracts;
+
+    }
+
     /**
      * Form for editing customer details
      * @access private
@@ -1649,6 +1698,80 @@ ORDER BY cus_name ASC  ";
                 )
             );
 
+
+        $dbeJRenContract = new DBEJRenContract($this);
+        $dbeJRenContract->getRowsByCustomerID($this->getCustomerID());
+        // broadband
+        $dbeJRenBroadband = new DBEJRenBroadband($this);
+        $dbeJRenBroadband->getRowsByCustomerID($this->getCustomerID());
+// Hosting
+        $dbeJRenHosting = new DBEJRenHosting($this);
+        $dbeJRenHosting->getRowsByCustomerID($this->getCustomerID());
+
+        $contracts = array_merge(
+            [],
+            $this->extractValidContracts($dbeJRenContract),
+            $this->extractValidContracts($dbeJRenBroadband),
+            $this->extractValidContracts($dbeJRenHosting)
+        );
+
+
+        uasort(
+            $contracts,
+            function ($a,
+                      $b
+            ) {
+                if (strcmp(
+                        $a['itemTypeDescription'],
+                        $b['itemTypeDescription']
+                    ) === 0) {
+                    return strcmp(
+                        $a['itemDescription'],
+                        $b['itemDescription']
+                    );
+                }
+                return strcmp(
+                    $a['itemTypeDescription'],
+                    $b['itemTypeDescription']
+                );
+            }
+        );
+        $this->template->set_block(
+            'CustomerEdit',
+            'toSignContractsBlock',
+            'toSignContracts'
+        );
+        $lastContractType = null;
+        foreach ($contracts as $contract) {
+
+            if ($contract['itemTypeDescription'] != $lastContractType) {
+                if ($lastContractType) {
+                    $optGroupClose = '</optgroup>';
+                } else {
+                    $optGroupClose = '';
+                }
+
+                $optGroupOpen = '<optgroup label="' . $contract['itemTypeDescription'] . '">';
+            } else {
+                $optGroupOpen = '';
+                $optGroupClose = '';
+            }
+            $lastContractType = $contract['itemTypeDescription'];
+
+            $this->template->set_var(
+                array(
+                    'toSignContractID'   => $contract['customerItemID'],
+                    'toSignContractName' => $contract['itemDescription'],
+                    'optGroupOpen'       => $optGroupOpen,
+                    'optGroupClose'      => $optGroupClose
+                )
+            );
+            $this->template->parse(
+                'toSignContracts',
+                'toSignContractsBlock',
+                true
+            );
+        }
         if ($_SESSION['save_page']) {
             $cancelURL = $_SESSION['save_page'];
         } else {